+++ OT282-42.py
@@ -1,0 +1,13 @@
+from datetime import datetime, timedelta
+import logging
+from airflow import DAG
+from airflow.operators.python_operator import PythonOperator
+
+
+logging.basicConfig(
+    format= '%(asctime)s - %(name)s - %(message)s',
+    level= logging.INFO,
+    datefmt='%Y-%m-%d'
+)
+
+logger = logging.getLogger('Dag logear eventos Universidades')
